Add method for handling integer properties.
git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/sling/trunk@660537 13f79535-47bb-0310-9956-ffa450edef68

+++ commons/osgi/src/main/java/org/apache/sling/commons/osgi/OsgiUtil.java
@@ -90,7 +90,7 @@ public class OsgiUtil {
      * the property is not an <code>Integer</code> and cannot be converted to
      * an <code>Integer</code> from the property's string value.
      */
-    public static long toInteger(Object propValue, int defaultValue) {
+    public static int toInteger(Object propValue, int defaultValue) {
         propValue = toObject(propValue);
         if (propValue instanceof Integer) {
             return (Integer) propValue;
